test: merge test files and pull it up
